+++ nextjs-dashboard/app/dashboard/(overview)/loading.tsx
@@ -1,0 +1,6 @@
+import DashboardSkeleton from "@/app/ui/skeletons";
+
+export default function Loading() {
+	return <DashboardSkeleton />;
+}
+
