+++ nextjs-dashboard/app/lib/data.ts
@@ -13,16 +13,7 @@ const sql = postgres(process.env.POSTGRES_URL!, { ssl: "require" });
 
 export async function fetchRevenue() {
 	try {
-		// Artificially delay a response for demo purposes.
-		// Don't do this in production :)
-
-		console.log("Fetching revenue data...");
-		await new Promise((resolve) => setTimeout(resolve, 3000));
-
 		const data = await sql<Revenue[]>`SELECT * FROM revenue`;
-
-		console.log("Data fetch completed after 3 seconds.");
-
 		return data;
 	} catch (error) {
 		console.error("Database Error:", error);
