+++ nextjs-dashboard/app/ui/dashboard/cards.tsx
@@ -1,58 +1,76 @@
 import {
-  BanknotesIcon,
-  ClockIcon,
-  UserGroupIcon,
-  InboxIcon,
-} from '@heroicons/react/24/outline';
-import { lusitana } from '@/app/ui/fonts';
+	BanknotesIcon,
+	ClockIcon,
+	UserGroupIcon,
+	InboxIcon,
+} from "@heroicons/react/24/outline";
+import { lusitana } from "@/app/ui/fonts";
+import { fetchCardData } from "@/app/lib/data";
 
 const iconMap = {
-  collected: BanknotesIcon,
-  customers: UserGroupIcon,
-  pending: ClockIcon,
-  invoices: InboxIcon,
+	collected: BanknotesIcon,
+	customers: UserGroupIcon,
+	pending: ClockIcon,
+	invoices: InboxIcon,
 };
 
 export default async function CardWrapper() {
-  return (
-    <>
-      {/* NOTE: Uncomment this code in Chapter 9 */}
+	const {
+		totalPaidInvoices,
+		totalPendingInvoices,
+		numberOfInvoices,
+		numberOfCustomers,
+	} = await fetchCardData();
 
-      {/* <Card title="Collected" value={totalPaidInvoices} type="collected" />
-      <Card title="Pending" value={totalPendingInvoices} type="pending" />
-      <Card title="Total Invoices" value={numberOfInvoices} type="invoices" />
-      <Card
-        title="Total Customers"
-        value={numberOfCustomers}
-        type="customers"
-      /> */}
-    </>
-  );
+	return (
+		<>
+			<Card
+				title="Collected"
+				value={totalPaidInvoices}
+				type="collected"
+			/>
+			<Card
+				title="Pending"
+				value={totalPendingInvoices}
+				type="pending"
+			/>
+			<Card
+				title="Total Invoices"
+				value={numberOfInvoices}
+				type="invoices"
+			/>
+			<Card
+				title="Total Customers"
+				value={numberOfCustomers}
+				type="customers"
+			/>
+		</>
+	);
 }
 
 export function Card({
-  title,
-  value,
-  type,
+	title,
+	value,
+	type,
 }: {
-  title: string;
-  value: number | string;
-  type: 'invoices' | 'customers' | 'pending' | 'collected';
+	title: string;
+	value: number | string;
+	type: "invoices" | "customers" | "pending" | "collected";
 }) {
-  const Icon = iconMap[type];
+	const Icon = iconMap[type];
 
-  return (
-    <div className="rounded-xl bg-gray-50 p-2 shadow-sm">
-      <div className="flex p-4">
-        {Icon ? <Icon className="h-5 w-5 text-gray-700" /> : null}
-        <h3 className="ml-2 text-sm font-medium">{title}</h3>
-      </div>
-      <p
-        className={`${lusitana.className}
-          truncate rounded-xl bg-white px-4 py-8 text-center text-2xl`}
-      >
-        {value}
-      </p>
-    </div>
-  );
+	return (
+		<div className="rounded-xl bg-gray-50 p-2 shadow-sm">
+			<div className="flex p-4">
+				{Icon ? <Icon className="h-5 w-5 text-gray-700" /> : null}
+				<h3 className="ml-2 text-sm font-medium">{title}</h3>
+			</div>
+			<p
+				className={`${lusitana.className}
+          truncate rounded-xl bg-white px-4 py-8 text-center text-2xl`}>
+				{value}
+			</p>
+		</div>
+	);
 }
+
