+++ nextjs-dashboard/app/ui/dashboard/revenue-chart.tsx
@@ -1,7 +1,7 @@
 import { generateYAxis } from "@/app/lib/utils";
 import { CalendarIcon } from "@heroicons/react/24/outline";
 import { lusitana } from "@/app/ui/fonts";
-import { Revenue } from "@/app/lib/definitions";
+import { fetchRevenue } from "@/app/lib/data";
 
 // This component is representational only.
 // For data visualization UI, check out:
@@ -9,11 +9,8 @@ import { Revenue } from "@/app/lib/definitions";
 // https://www.chartjs.org/
 // https://airbnb.io/visx/
 
-export default async function RevenueChart({
-	revenue,
-}: {
-	revenue: Revenue[];
-}) {
+export default async function RevenueChart() {
+	const revenue = await fetchRevenue();
 	const chartHeight = 350;
 
 	const { yAxisLabels, topLabel } = generateYAxis(revenue);
